fix: Improve JSON sorting and extraction robustness
- Cast severity to INTEGER for proper numeric sorting in JSON output
- Simplify JSON extraction to use direct grep without complex pipes
- Addresses PR review feedback from copilot-pull-request-reviewer

+++ service.data.impl/src/main/resources/schema-health-check.sql
@@ -743,7 +743,7 @@ SELECT json_build_object(
         ) cat
     ),
     'issues', (
-        SELECT json_agg(issue_obj ORDER BY issue_obj->>'severity' DESC)
+        SELECT json_agg(issue_obj ORDER BY (issue_obj->>'severity')::INTEGER DESC)
         FROM (
             SELECT json_build_object(
                 'category', category,
